==> notebooks/de/2_baseline.ipynb <==
import torch
import torch.utils.data as tud
import torch.nn as nn
import pickle
from tqdm.notebook import tqdm
import datetime
import matplotlib.pyplot as plt
import importlib
import sys
import pandas as pd

!pwd

sys.path.append("/home/jarobyte/guemes/lib")

from pytorch_decoding.seq2seq import Transformer

scratch = "/home/jarobyte/scratch/guemes/icdar/de/"

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
device

char2i = pickle.load(open(scratch + "data/char2i.pkl", "rb"))

i2char = pickle.load(open(scratch + "data/i2char.pkl", "rb"))

# data

train_size = 1500000

dev_size = 1000000

# train_size = 1000

# dev_size = 100

## train

train_source = torch.load(scratch + "data/train_source.pt")[:train_size].to(device)
train_source.shape

train_target = torch.load(scratch + "data/train_target.pt")[:train_size].to(device)
train_target.shape

dev_source = torch.load(scratch + "data/dev_source.pt")[:dev_size].to(device)
dev_source.shape

dev_target = torch.load(scratch + "data/dev_target.pt")[:dev_size].to(device)
dev_target.shape

# model

model = Transformer(char2i, 
                    i2char, 
                    max_sequence_length = 110,
                    embedding_dimension = 512,
                    feedforward_dimension = 2048,
                    attention_heads = 8,
                    encoder_layers = 4,
                    decoder_layers = 4)
model.to(device)

log = model.fit(train_source, 
                train_target, 
                dev_source, 
                dev_target, 
                epochs = 1, 
                progress_bar = 2, 
                learning_rate = 10**-4)

idx, probs = model.predict(dev_source)

model.tensor2text(idx)

model.tensor2text(train_target[:10])

model.tensor2text(train_source[:10])
