==> notebooks/de/6_evaluation.ipynb <==
import importlib
import pickle
import pandas as pd
import torch
import sys
sys.path.append("/home/jarobyte/guemes/lib")
from pytorch_decoding import seq2seq
import ocr_correction
from timeit import default_timer as t
from tqdm.notebook import tqdm
from ocr_correction import *

importlib.reload(ocr_correction)
language = "de"
model_id = "22694260_9"

scratch = f"/home/jarobyte/scratch/guemes/icdar/{language}/"

print(f"Evaluating {language}...")
data = pd.read_pickle(f"/home/jarobyte/scratch/guemes/icdar/{language}/data/test.pkl")
device = torch.device("cuda")
model = seq2seq.load_architecture(scratch + f"baseline/models/{model_id}.arch")
model.to(device)
model.eval()
model.load_state_dict(torch.load(scratch + f"baseline/checkpoints/{model_id}.pt"))
with open(scratch + "data/vocabulary.pkl", "rb") as file:
    vocabulary = pickle.load(file)
len(vocabulary)

raw = data.ocr_to_input
gs = data.gs_aligned
document_progress_bar = 0
window_size = 50
metrics = []
old = levenshtein(reference = gs, hypothesis = raw).cer.mean()

start = t()
corrections = [correct_by_disjoint_window(s, 
                                          model, 
                                          vocabulary, 
                                          document_progress_bar = 0, 
                                          window_size = window_size) 
               for s in tqdm(raw)]
metrics.append({"window":"disjoint", 
                "decoding":
                "greedy",
                "window_size":window_size * 2,
                "inference_seconds":t() - start,
                "cer_before":old,
                "cer_after":levenshtein(gs, corrections).cer.mean()})

start = t()
corrections = [correct_by_disjoint_window(s, 
                                          model, 
                                          vocabulary, 
                                          document_progress_bar = document_progress_bar, 
                                          window_size = window_size * 2) 
               for s in raw]
metrics.append({"window":"disjoint", 
                "decoding":"greedy",
                "window_size":window_size,
                "inference_seconds":t() - start,
                "cer_before":old,
                "cer_after":levenshtein(gs, corrections).cer.mean()})

start = t()
corrections = [correct_by_disjoint_window(s, 
                                          model, 
                                          vocabulary, 
                                          decoding_method = "beam_search", 
                                          document_progress_bar = document_progress_bar, 
                                          window_size = window_size) 
               for s in tqdm(raw)]
metrics.append({"window":"disjoint", 
                "decoding":"beam", 
                "window_size":window_size * 2,
                "inference_seconds":t() - start,
                "cer_before":old,
                "cer_after":levenshtein(gs, corrections).cer.mean()})
start = t()
corrections = [correct_by_disjoint_window(s, 
                                          model, 
                                          vocabulary, 
                                          decoding_method = "beam_search", 
                                          document_progress_bar = document_progress_bar, 
                                          window_size = window_size * 2) 
               for s in tqdm(raw)]
metrics.append({"window":"disjoint", 
                "decoding":"beam", 
                "window_size":window_size,
                "inference_seconds":t() - start,
                "cer_before":old,
                "cer_after":levenshtein(gs, corrections).cer.mean()})

start = t()
corrections = [correct_by_sliding_window(s, model, vocabulary, 
                                         weighting = uniform,
                                         document_progress_bar = document_progress_bar,
                                         window_size = window_size)[1] 
               for s in tqdm(raw)]
metrics.append({"window":"sliding", 
                "decoding":"greedy", 
                "weighting":"uniform",
                "window_size":window_size,
                "inference_seconds":t() - start,
                "cer_before":old,
                "cer_after":levenshtein(gs, corrections).cer.mean()})

start = t()
corrections = [correct_by_sliding_window(s, model, vocabulary, 
                                         decoding_method = "beam_search", 
                                         weighting = uniform, 
                                         document_progress_bar = document_progress_bar,
                                         window_size = window_size)[1] 
               for s in tqdm(raw)]
metrics.append({"window":"sliding", 
                "decoding":"beam", 
                "weighting":"uniform",
                "window_size":window_size,
                "inference_seconds":t() - start,
                "cer_before":old,
                "cer_after":levenshtein(gs, corrections).cer.mean()})

pd.DataFrame(metrics).assign(improvement = lambda df: 100 * (1 - df.cer_after / df.cer_before))
